Rename Component folder/namespace
